register update

+++ WebShopManagement/Webshop/Webshop/DataContracts/CreateProductModel.cs
@@ -6,5 +6,5 @@ public class CreateProductModel
         public required string Description { get; set; }
         public required decimal Price { get; set; }
         public required int Stock { get; set; }
-        public int? DiscountPercentage { get; set; }
+        public decimal? DiscountPercentage { get; set; }
 }
