+++ WebShopManagement/Webshop/Webshop/DataContracts/ProductViewModel.cs
@@ -7,7 +7,7 @@ public class ProductViewModel
     public required string Description { get; set; }
     public required decimal Price { get; set; }
     public required int Stock { get; set; }
-    public int? DiscountPercentage { get; set; }
+    public decimal? DiscountPercentage { get; set; }
     public decimal DiscountedPrice => 
     DiscountPercentage.HasValue ? Price * (1 - DiscountPercentage.Value / 100m) : Price;
