+++ WebShopManagement/Webshop/Webshop/Models/Product.cs
@@ -9,7 +9,7 @@ public class Product
     public required string Description { get; set; }
     public required decimal Price { get; set; }
     public required int Stock { get; set; }
-    public int? DiscountPercentage { get; set; }
+    public decimal? DiscountPercentage { get; set; }
     
     [NotMapped]
     public decimal DiscountedPrice =>
